pyproject.toml
[project]
name = "hide-and-seek"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml==6.0.2",
    "numpy==2.2.5",
    "matplotlib==3.10.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "pytest-cov>=4.1.0",
    "ruff==0.11.4",
    "mypy>=1.3.0",
    "pre-commit>=3.3.2",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"